add cron job
- send email every weekday
- closes #1
