+++ package.json
@@ -29,6 +29,7 @@
   "dependencies": {
     "bluebird": "^3.3.5",
     "body-parser": "^1.15.0",
+    "cron": "^1.1.0",
     "express": "^4.13.3",
     "express-handlebars": "^3.0.0",
     "express-sslify": "^1.0.1",
